+++ src/frequenz/sdk/microgrid/_data_pipeline.py
@@ -24,6 +24,7 @@ from ..timeseries._base_types import PoolType
 from ..timeseries._grid_frequency import GridFrequency
 from ..timeseries.grid import Grid
 from ..timeseries.grid import get as get_grid
+from ..timeseries.grid import initialize as initialize_grid
 from . import connection_manager
 from .component import ComponentCategory
 
@@ -115,6 +116,7 @@ class _DataPipeline:  # pylint: disable=too-many-instance-attributes
         self._power_managing_actor: _power_managing.PowerManagingActor | None = None
 
         self._logical_meter: LogicalMeter | None = None
+        self._grid: Grid | None = None
         self._ev_charger_pools: dict[frozenset[int], EVChargerPool] = {}
         self._battery_pools: dict[frozenset[int], BatteryPoolReferenceStore] = {}
         self._frequency_pool: dict[int, GridFrequency] = {}
@@ -189,9 +191,7 @@ class _DataPipeline:  # pylint: disable=too-many-instance-attributes
             )
         return self._ev_charger_pools[key]
 
-    def grid(
-        self,
-    ) -> Grid | None:
+    def grid(self) -> Grid:
         """Return the grid instance.
 
         If a Grid instance doesn't exist, a new one is created and returned.
@@ -199,7 +199,14 @@ class _DataPipeline:  # pylint: disable=too-many-instance-attributes
         Returns:
             A Grid instance.
         """
-        return get_grid()
+        if self._grid is None:
+            initialize_grid(
+                channel_registry=self._channel_registry,
+                resampler_subscription_sender=self._resampling_request_sender(),
+            )
+            self._grid = get_grid()
+
+        return self._grid
 
     def battery_pool(
         self,
@@ -470,7 +477,7 @@ def battery_pool(
     return _get().battery_pool(battery_ids, name, priority)
 
 
-def grid() -> Grid | None:
+def grid() -> Grid:
     """Return the grid instance.
 
     Returns: