+++ src/frequenz/sdk/timeseries/logical_meter/_logical_meter.py
@@ -10,14 +10,12 @@ from frequenz.channels import Sender
 
 from ...actor import ChannelRegistry, ComponentMetricRequest
 from ...microgrid.component import ComponentMetricId
-from .._quantities import Current, Power, Quantity
-from ..formula_engine import FormulaEngine, FormulaEngine3Phase
+from .._quantities import Power, Quantity
+from ..formula_engine import FormulaEngine
 from ..formula_engine._formula_engine_pool import FormulaEnginePool
 from ..formula_engine._formula_generators import (
     CHPPowerFormula,
     ConsumerPowerFormula,
-    GridCurrentFormula,
-    GridPowerFormula,
     ProducerPowerFormula,
     PVPowerFormula,
 )
@@ -49,17 +47,17 @@ class LogicalMeter:
         )
 
         logical_meter = microgrid.logical_meter()
+        grid = microgrid.grid()
 
         # Get a receiver for a builtin formula
-        grid_power_recv = logical_meter.grid_power.new_receiver()
-        async for grid_power_sample in grid_power_recv:
-            print(grid_power_sample)
+        consumer_power_recv = logical_meter.consumer_power.new_receiver()
+        async for consumer_power_sample in consumer_power_recv:
+            print(consumer_power_sample)
 
         # or compose formulas to create a new formula
         net_power_recv = (
             (
-                logical_meter.grid_power
-                - logical_meter.pv_power
+                grid.power - logical_meter.pv_power
             )
             .build("net_power")
             .new_receiver()
@@ -127,50 +125,6 @@ class LogicalMeter:
         return self._formula_pool.from_string(
             formula, component_metric_id, nones_are_zeros=nones_are_zeros
         )
-
-    @property
-    def grid_power(self) -> FormulaEngine[Power]:
-        """Fetch the grid power for the microgrid.
-
-        This formula produces values that are in the Passive Sign Convention (PSC).
-
-        If a formula engine to calculate grid power is not already running, it will be
-        started.
-
-        A receiver from the formula engine can be created using the `new_receiver`
-        method.
-
-        Returns:
-            A FormulaEngine that will calculate and stream grid power.
-        """
-        engine = self._formula_pool.from_power_formula_generator(
-            "grid_power",
-            GridPowerFormula,
-        )
-        assert isinstance(engine, FormulaEngine)
-        return engine
-
-    @property
-    def grid_current(self) -> FormulaEngine3Phase[Current]:
-        """Fetch the grid power for the microgrid.
-
-        This formula produces values that are in the Passive Sign Convention (PSC).
-
-        If a formula engine to calculate grid current is not already running, it will be
-        started.
-
-        A receiver from the formula engine can be created using the `new_receiver`
-        method.
-
-        Returns:
-            A FormulaEngine that will calculate and stream grid current.
-        """
-        engine = self._formula_pool.from_3_phase_current_formula_generator(
-            "grid_current",
-            GridCurrentFormula,
-        )
-        assert isinstance(engine, FormulaEngine3Phase)
-        return engine
 
     @property
     def consumer_power(self) -> FormulaEngine[Power]: